add files
Density-based clustering is one of the most popular paradigms of existing clustering approaches, most approaches of this kind, such as DBSCAN, recognize clusters of data characterized by a ﬁxed scanning radius. However, some ﬂaws are caused by the ﬁxed scanning radius, e.g., the determination of a proper scanning radius is nontrivial. In order to solve these problems, we revise DBSCAN, Meanshift, DPeak, etc. based on two new features, i.e., effective radius and density heat (DHeat). Generally, we name these revised clustering algorithms as DHeat. The underlying idea is based on two assumptions: 1) the existence of clusters is raised by the nonuniformity of data distribution, and the density of one data point within its r-neighborhood is proportional to the volume of the neighborhood provided the density distribution is uniform and 2) each cluster can be divided into different density layers, such as edges, shallow inner, deep inner, etc.; the deeper inner of a point locates, the higher density of that point. The experiments conducted on various test cases show that the advantage of DHeat lies in its good performance and the self-adapting scanning radius.